old-tests/submission/W3C/domhtml/HTMLObjectElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-object-align" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<object align="middle" data="/support/html/sample.html" height="80" type="text/html" width="550"></object>

<script>
var domtest = async_test("The align IDL attribute of the object element must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var valign;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
      testNode = nodeList.item(0);
      valign = testNode.align;

      assert_equals(valign,"middle");
       
    domtest.done();
}));

</script>
<p>The align IDL attribute of the object element must reflect the respective content attribute of the same name.
</p>
<div id="log"></div>
</body></html>
